+++ lib/rpush/deprecatable.rb
@@ -14,8 +14,7 @@ module Rpush
         warning << " #{msg}" if msg
         class_eval(<<-RUBY, __FILE__, __LINE__)
           def #{method_name}(*args, &blk)
-            trace = "\n\nCALLED FROM:\n" + caller.join("\n")
-            Rpush::Deprecation.warn(#{warning.inspect} + trace)
+            Rpush::Deprecation.warn_with_backtrace(#{warning.inspect})
             #{method_name_as_var}_without_warning(*args, &blk)
           end
         RUBY