+++ lib/rpush/push.rb
@@ -2,6 +2,11 @@ module Rpush
   def self.push(options = {})
     require 'rpush/daemon'
 
+    if !options.empty?
+      warning = "Passing configuration options directly to Rpush.push is deprecated and will be removed from Rpush 2.5.0. Please setup configuration using Rpush.configure { |config| ... } before calling push."
+      Rpush::Deprecation.warn_with_backtrace(warning)
+    end
+
     config = Rpush::ConfigurationWithoutDefaults.new
     options.each { |k, v| config.send("#{k}=", v) }
     config.push = true