+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Support for MongoDB (using Mongoid).
   * config.feedback_poll is now deprecated, use config.apns.feedback_receiver.frequency instead.
   * Add config.apns.feedback_receiver.enabled to optionally enabled the APNs feedback receiver (#129).
+  * Passing configuration options directly to Rpush.embed and Rpush.push is now deprecated.
 
   Bug fixes:
   * Fix setting the log level when using Rails 4+ or without Rails (#124).